pyproject.toml
[project]
name = "testtaskbs01"
version = "0.1.0"
description = "Senior-level API automation framework for httpbin.org (pytest)."
requires-python = ">=3.11"
dependencies = [
  "pytest>=8.0",
  "requests>=2.32",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "PyYAML>=6.0",
  "faker>=26.0",
  "tenacity>=9.0",
  "allure-pytest>=2.13.5",
  "pytest-html>=4.1.1",
  "prometheus-client>=0.20.0",
]

[tool.pytest.ini_options]
addopts = "-q --disable-warnings --maxfail=1"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
branch = true
source = ["framework", "tests"]

[tool.coverage.report]
show_missing = true
skip_covered = true
